[refactor] 디렉토리 구조, 엔티티 리팩토링
