--- src/main/java/com/ongi/backend/common/error/ErrorCase.java
@@ -1,7 +1,0 @@
-package com.ongi.backend.common.error;
-
-public interface ErrorCase  {
-    Integer getHttpStatusCode();
-    Integer getErrorCode();
-    String getMessage();
-}
