+++ src/main/java/com/ongi/backend/common/exception/ErrorCase.java
@@ -1,0 +1,7 @@
+package com.ongi.backend.common.exception;
+
+public interface ErrorCase  {
+    Integer getHttpStatusCode();
+    Integer getErrorCode();
+    String getMessage();
+}
