+++ src/main/java/com/ongi/backend/domain/caregiver/dto/request/WorkTimeRequestDto.java
@@ -1,0 +1,24 @@
+package com.ongi.backend.domain.caregiver.dto.request;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.DayOfWeek;
+import java.time.LocalTime;
+
+@Getter
+@NoArgsConstructor
+public class WorkTimeRequestDto {
+
+    private DayOfWeek dayOfWeek;
+    private LocalTime startTime;
+    private LocalTime endTime;
+
+    @Builder
+    public WorkTimeRequestDto(DayOfWeek dayOfWeek, LocalTime startTime, LocalTime endTime) {
+        this.dayOfWeek = dayOfWeek;
+        this.startTime = startTime;
+        this.endTime = endTime;
+    }
+}
