+++ src/main/java/com/ongi/backend/domain/caregiver/exception/CaregiverException.java
@@ -1,0 +1,4 @@
+package com.ongi.backend.domain.caregiver.exception;
+
+public class CaregiverException {
+}
