+++ src/main/java/com/ongi/backend/domain/caregiver/repository/CaregiverLicenseRepository.java
@@ -1,0 +1,7 @@
+package com.ongi.backend.domain.caregiver.repository;
+
+import com.ongi.backend.domain.caregiver.entity.CaregiverLicense;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CaregiverLicenseRepository extends JpaRepository<CaregiverLicense, Long> {
+}
