+++ src/main/java/com/ongi/backend/domain/caregiver/repository/CaregiverRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.ongi.backend.domain.caregiver.repository;
+
+public interface CaregiverRepositoryCustom {
+}
