--- src/main/java/com/ongi/backend/repository/caregiver/CaregiverRepository.java
@@ -1,7 +1,0 @@
-package com.ongi.backend.repository.caregiver;
-
-import com.ongi.backend.entity.caregiver.Caregiver;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CaregiverRepository extends JpaRepository<Caregiver, Long>, CaregiverRepositoryCustom {
-}
