--- src/main/java/com/ongi/backend/repository/caregiver/CaregiverRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.ongi.backend.repository.caregiver;
-
-public interface CaregiverRepositoryCustom {
-}
