--- src/main/java/com/ongi/backend/repository/caregiver/WorkRegionRepository.java
@@ -1,9 +1,0 @@
-package com.ongi.backend.repository.caregiver;
-
-import com.ongi.backend.entity.caregiver.CaregiverWorkCondition;
-import com.ongi.backend.entity.caregiver.WorkRegion;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface WorkRegionRepository extends JpaRepository<WorkRegion, Long> {
-    void deleteByWorkCondition(CaregiverWorkCondition workCondition);
-}
