--- src/main/java/com/ongi/backend/repository/caregiver/WorkTimeRepository.java
@@ -1,9 +1,0 @@
-package com.ongi.backend.repository.caregiver;
-
-import com.ongi.backend.entity.caregiver.CaregiverWorkCondition;
-import com.ongi.backend.entity.caregiver.WorkTime;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface WorkTimeRepository extends JpaRepository<WorkTime, Long> {
-    void deleteByWorkCondition(CaregiverWorkCondition workCondition);
-}
